Added the current code we have tested
As of now, it will generate an emoji mosaic of the emoji "1.png" and it only uses one emoji to generate that mosaic. Also, in order to make it appear like a mosaic, the emoji has a filter on it that will make it the color of the corresponding sub section of the original image.